feat: split worker_service.py into worker, worker_stash, and worker_service
